% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_svg.R
\name{clean_svg}
\alias{clean_svg}
\title{Cleans an svg file (created by ggsave) for use in vector design software}
\usage{
clean_svg(file_path)
}
\arguments{
\item{file_path}{the location of the svg file to be cleaned}
}
\value{
nothing - executed for side effects (edits on an svg file)
}
\description{
Cleans an svg file (created by ggsave) for use in vector design software
}
\examples{

svg_name <- "test.svg" # define file name

# create plot
ggplot(mtcars, aes(mpg, wt)) +
geom_point()

# save plot as svg file
ggsave(svg_name)
# clean up the svg file
clean_svg(svg_name)

}
