% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_design.R
\name{geom_design}
\alias{geom_design}
\title{Design}
\usage{
geom_design(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{type}{specifies the format of the units to plot; one of
["square"], ["hex"], or ["hex_tri"]. By default
this is ["square"]}

\item{seg}{ignored unless [type = "hex_tri"]. Specifies the
`side` of the hex unit to fill; one of [1] or [2] (by default = [1])}
}
\description{
Design
}
\details{
Used to plot illustrative figures of experimental design structures
}
\examples{
crd <- data.frame(plots = 1:6, trt = c("C", "B", "B", "A", "A", "C"))

ggplot(crd, aes(plots = plots, fill = trt)) + geom_design()

ggplot(crd, aes(plots = plots, fill = trt)) +
geom_design(type = "hex") + scale_fill_brewer(palette = "Dark2")
}
