% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{get_colours}
\alias{get_colours}
\title{Get the first n colours in the ggdesign palette}
\usage{
get_colours(n = length(colours))
}
\arguments{
\item{n}{number of colours requested from the palette}
}
\value{
a vector of colours (as hexcodes)
}
\description{
Get the first n colours in the ggdesign palette
}
\examples{
get_colours(2)

}
