% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_edit.R
\name{theme_edit}
\alias{theme_edit}
\title{Style a ggplot chart to make it easier to edit in vector design software}
\usage{
theme_edit()
}
\value{
an object with classes gg and theme which applies styling to a ggplot
object
}
\description{
Style a ggplot chart to make it easier to edit in vector design software
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(mpg, wt)) +
ggplot2::geom_point() +
theme_edit()

}
