% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_to_size.R
\name{theme_to_size}
\alias{theme_to_size}
\title{Resizes chart structure (lines width and text size) based on the size of the
chart}
\usage{
theme_to_size(width = 90, height = 60, units = "mm")
}
\arguments{
\item{width}{Width of plot to be exported}

\item{height}{Height of plot to be exported}

\item{units}{Units for height and width ("in", "cm", "mm", or "px").}
}
\value{
an object with classes gg and theme which applies theme styling to a
ggplot object
}
\description{
Resizes chart structure (lines width and text size) based on the size of the
chart
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(mpg, wt)) +
  ggplot2::geom_point() +
  theme_to_size(width = 90, height = 60, units = "mm")

}
