% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_axis.R
\name{clean_axis}
\alias{clean_axis}
\title{Cleans up the axis on a ggplot2 chart (ahead of export for editing in vector
design software)}
\usage{
clean_axis(coord_system = ggplot2::coord_cartesian)
}
\arguments{
\item{coord_system}{the ggplot2 coordinate system used in the chart (e.g.
ggplot2::coord_equal)}
}
\value{
an object with the following classes to be added to chart using the
'+' operator. Expected classes: CoordCartesian, Coord, ggproto, gg. This
cleans up the axis.
}
\description{
Cleans up the axis on a ggplot2 chart (ahead of export for editing in vector
design software)
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(mpg, wt)) +
ggplot2::geom_point() +
clean_axis()

}
